Refactor sentence fusion helpers into dedicated module
